+++ backend/src/main/java/com/multi/mariage/review/service/ReviewFindService.java
@@ -20,9 +20,9 @@ import com.multi.mariage.review.dto.response.ProductReviewsResponse;
 import com.multi.mariage.review.exception.ReviewErrorCode;
 import com.multi.mariage.review.exception.ReviewException;
 import com.multi.mariage.review.vo.ProductReviewVO;
-import com.multi.mariage.review.vo.myreview.MyProductReviewVO;
-import com.multi.mariage.review.vo.myreview.MyReviewVO;
-import com.multi.mariage.review.vo.myreview.ReviewContentVO;
+import com.multi.mariage.review.vo.my.review.MyProductReviewVO;
+import com.multi.mariage.review.vo.my.review.MyReviewVO;
+import com.multi.mariage.review.vo.my.review.ReviewContentVO;
 import com.multi.mariage.review.vo.product.ProductReviewContentVO;
 import com.multi.mariage.review.vo.product.ProductReviewFoodVO;
 import com.multi.mariage.review.vo.product.ProductReviewLikeVO;
